Renamed package for entire project to com.capstone.event_finder
