+++ app/build.gradle
@@ -10,7 +10,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.event_finder_capstone"
+        applicationId "com.capstone.event_finder"
         minSdk 29
         targetSdk 32
         versionCode 1
