+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.event_finder_capstone">
+    package="com.capstone.event_finder">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -12,7 +12,7 @@
     </queries>
 
     <application
-        android:name=".network.ParseApplication"
+        android:name="com.capstone.event_finder.network.ParseApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -25,26 +25,26 @@
 
         <provider
             android:name="androidx.core.content.FileProvider"
-            android:authorities="com.codepath.photoprovider"
+            android:authorities="com.codepath.event_photo_provider"
             android:exported="false"
             android:grantUriPermissions="true">
             <meta-data
                 android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/photoprovider" />
+                android:resource="@xml/event_photo_provider" />
         </provider>
 
         <activity
-            android:name=".activities.LoginActivity"
+            android:name="com.capstone.event_finder.activities.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.MainActivity" />
-        <activity android:name=".activities.SignUpActivity" />
-        <activity android:name=".activities.EventDetailsActivity" />
-        <activity android:name=".activities.PhotoAlbumActivity" />
-        <activity android:name=".activities.AddPhotoActivity" />
+        <activity android:name="com.capstone.event_finder.activities.MainActivity" />
+        <activity android:name="com.capstone.event_finder.activities.SignUpActivity" />
+        <activity android:name="com.capstone.event_finder.activities.EventDetailsActivity" />
+        <activity android:name="com.capstone.event_finder.activities.PhotoAlbumActivity" />
+        <activity android:name="com.capstone.event_finder.activities.AddPhotoActivity" />
     </application>
 </manifest>
