+++ app/src/main/java/com/capstone/event_finder/models/Event.java
@@ -1,0 +1,238 @@
+package com.capstone.event_finder.models;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+import org.parceler.Parcel;
+
+import javax.annotation.Generated;
+
+@Parcel(analyze = {Event.class})
+@Generated("jsonschema2pojo")
+public class Event {
+
+    @SerializedName("attending_count")
+    @Expose
+    private int attendingCount;
+    @SerializedName("category")
+    @Expose
+    private String category;
+    @SerializedName("cost")
+    @Expose
+    private String cost;
+    @SerializedName("cost_max")
+    @Expose
+    private String costMax;
+    @SerializedName("description")
+    @Expose
+    private String description;
+    @SerializedName("event_site_url")
+    @Expose
+    private String eventSiteUrl;
+    @SerializedName("id")
+    @Expose
+    private String id;
+    @SerializedName("image_url")
+    @Expose
+    private String imageUrl;
+    @SerializedName("interested_count")
+    @Expose
+    private int interestedCount;
+    @SerializedName("is_canceled")
+    @Expose
+    private boolean isCanceled;
+    @SerializedName("is_free")
+    @Expose
+    private boolean isFree;
+    @SerializedName("is_official")
+    @Expose
+    private boolean isOfficial;
+    @SerializedName("latitude")
+    @Expose
+    private float latitude;
+    @SerializedName("longitude")
+    @Expose
+    private float longitude;
+    @SerializedName("name")
+    @Expose
+    private String name;
+    @SerializedName("tickets_url")
+    @Expose
+    private String ticketsUrl;
+    @SerializedName("time_end")
+    @Expose
+    private String timeEnd;
+    @SerializedName("time_start")
+    @Expose
+    private String timeStart;
+    @SerializedName("location")
+    @Expose
+    private String location;
+    @SerializedName("business_id")
+    @Expose
+    private String businessId;
+
+    public Event() {
+    }
+
+    public int getAttendingCount() {
+        return attendingCount;
+    }
+
+    public void setAttendingCount(int attendingCount) {
+        this.attendingCount = attendingCount;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    public String getCost() {
+        return cost;
+    }
+
+    public void setCost(String cost) {
+        this.cost = (String) cost;
+    }
+
+    public String getCostMax() {
+        return costMax;
+    }
+
+    public void setCostMax(String costMax) {
+        this.costMax = (String) costMax;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getEventSiteUrl() {
+        return eventSiteUrl;
+    }
+
+    public void setEventSiteUrl(String eventSiteUrl) {
+        this.eventSiteUrl = eventSiteUrl;
+    }
+
+    public String getId() {
+        return (String) id;
+    }
+
+    public void setId(String id) {
+        this.id = (String) id;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public void setImageUrl(String imageUrl) {
+        this.imageUrl = imageUrl;
+    }
+
+    public int getInterestedCount() {
+        return interestedCount;
+    }
+
+    public void setInterestedCount(int interestedCount) {
+        this.interestedCount = interestedCount;
+    }
+
+    public boolean isIsCanceled() {
+        return isCanceled;
+    }
+
+    public void setIsCanceled(boolean isCanceled) {
+        this.isCanceled = isCanceled;
+    }
+
+    public boolean isIsFree() {
+        return isFree;
+    }
+
+    public void setIsFree(boolean isFree) {
+        this.isFree = isFree;
+    }
+
+    public boolean isIsOfficial() {
+        return isOfficial;
+    }
+
+    public void setIsOfficial(boolean isOfficial) {
+        this.isOfficial = isOfficial;
+    }
+
+    public float getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(float latitude) {
+        this.latitude = latitude;
+    }
+
+    public float getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(float longitude) {
+        this.longitude = longitude;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getTicketsUrl() {
+        return ticketsUrl;
+    }
+
+    public void setTicketsUrl(String ticketsUrl) {
+        this.ticketsUrl = ticketsUrl;
+    }
+
+    public String getTimeEnd() {
+        return timeEnd;
+    }
+
+    public void setTimeEnd(String timeEnd) {
+        this.timeEnd = timeEnd;
+    }
+
+    public String getTimeStart() {
+        return timeStart;
+    }
+
+    public void setTimeStart(String timeStart) {
+        this.timeStart = timeStart;
+    }
+
+    public String getLocation() {
+        return location.toString();
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+
+    public String getBusinessId() {
+        return businessId;
+    }
+
+    public void setBusinessId(String businessId) {
+        this.businessId = businessId;
+    }
+
+}
