+++ app/src/main/java/com/capstone/event_finder/utils/BitmapScaler.java
@@ -1,0 +1,10 @@
+package com.capstone.event_finder.utils;
+
+import android.graphics.Bitmap;
+
+public class BitmapScaler {
+    public static Bitmap scaleToFitWidth(Bitmap b, int width) {
+        float factor = width / (float) b.getWidth();
+        return Bitmap.createScaledBitmap(b, width, (int) (b.getHeight() * factor), true);
+    }
+}
