--- app/src/main/java/com/example/event_finder_capstone/adapters/PhotoAdapter.java
@@ -1,63 +1,0 @@
-package com.example.event_finder_capstone.adapters;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.bumptech.glide.Glide;
-import com.example.event_finder_capstone.R;
-import com.example.event_finder_capstone.models.Photo;
-import com.parse.ParseFile;
-
-import java.util.List;
-
-public class PhotoAdapter extends RecyclerView.Adapter<PhotoAdapter.ViewHolder> {
-
-    private final Context context;
-    private final List<Photo> photos;
-
-    public PhotoAdapter(Context context, List<Photo> photos) {
-        this.context = context;
-        this.photos = photos;
-    }
-
-    @NonNull
-    @Override
-    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(context).inflate(R.layout.item_photo, parent, false);
-        return new ViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        Photo post = photos.get(position);
-        holder.bind(post);
-    }
-
-    @Override
-    public int getItemCount() {
-        return photos.size();
-    }
-
-    public class ViewHolder extends RecyclerView.ViewHolder {
-
-        private final ImageView ivAlbumPhotoItem;
-
-        public ViewHolder(@NonNull View itemView) {
-            super(itemView);
-            ivAlbumPhotoItem = itemView.findViewById(R.id.ivAlbumPhotoItem);
-        }
-
-        public void bind(Photo post) {
-            ParseFile image = post.getImage();
-            if (image != null) {
-                Glide.with(context).load(image.getUrl()).into(ivAlbumPhotoItem);
-            }
-        }
-    }
-}
