--- app/src/main/java/com/example/event_finder_capstone/interfaces/EventEndpointsInterface.java
@@ -1,16 +1,0 @@
-package com.example.event_finder_capstone.interfaces;
-
-import com.google.gson.JsonObject;
-
-import retrofit2.Call;
-import retrofit2.http.GET;
-import retrofit2.http.Query;
-
-public interface EventEndpointsInterface {
-    @GET("events")
-    Call<JsonObject> getEvents(@Query("locale") String locale,
-                               @Query("limit") String limit,
-                               @Query("start_date") Long start_date,
-                               @Query("radius") Long radius,
-                               @Query("location") String location);
-}
