--- app/src/main/java/com/example/event_finder_capstone/models/Photo.java
@@ -1,47 +1,0 @@
-package com.example.event_finder_capstone.models;
-
-import com.parse.ParseClassName;
-import com.parse.ParseFile;
-import com.parse.ParseObject;
-import com.parse.ParseUser;
-
-@ParseClassName("Photo")
-public class Photo extends ParseObject {
-    public static final String KEY_DESCRIPTION = "description";
-    public static final String KEY_IMAGE = "image";
-    public static final String KEY_USER = "user";
-    public static final String KEY_EVENT_ID = "eventId";
-
-    public String getDescription() {
-        return getString(KEY_DESCRIPTION);
-    }
-
-    public void setDescription(String description) {
-        put(KEY_DESCRIPTION, description);
-    }
-
-    public ParseFile getImage() {
-        return getParseFile(KEY_IMAGE);
-    }
-
-    public void setImage(ParseFile parseFile) {
-        put(KEY_IMAGE, parseFile);
-    }
-
-    public ParseUser getUser() {
-        return getParseUser(KEY_USER);
-    }
-
-    public void setUser(ParseUser user) {
-        put(KEY_USER, user);
-    }
-
-    public String getEventId() {
-        return getString(KEY_EVENT_ID);
-    }
-
-    public void setEventId(String eventId) {
-        put(KEY_EVENT_ID, eventId);
-    }
-
-}
