--- app/src/main/java/com/example/event_finder_capstone/utils/BitmapScaler.java
@@ -1,10 +1,0 @@
-package com.example.event_finder_capstone.utils;
-
-import android.graphics.Bitmap;
-
-public class BitmapScaler {
-    public static Bitmap scaleToFitWidth(Bitmap b, int width) {
-        float factor = width / (float) b.getWidth();
-        return Bitmap.createScaledBitmap(b, width, (int) (b.getHeight() * factor), true);
-    }
-}
