+++ app/src/main/res/layout/activity_photo_album.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".activities.PhotoAlbumActivity">
+    tools:context="com.capstone.event_finder.activities.PhotoAlbumActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvPhotos"
