+++ app/src/main/res/layout/activity_signup.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".activities.MainActivity">
+    tools:context="com.capstone.event_finder.activities.MainActivity">
 
     <EditText
         android:id="@+id/etUsername"
