+++ app/src/main/res/layout/fragment_explore.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.ExploreFragment">
+    tools:context="com.capstone.event_finder.fragments.ExploreFragment">
 
     <TextView
         android:layout_width="match_parent"
