+++ app/src/main/res/layout/fragment_feed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.MainActivity">
+    tools:context="com.capstone.event_finder.activities.MainActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvEvents"
