+++ app/src/main/res/xml/event_photo_provider.xml
